Refactor unit tests into their own namespace
