+++ phpunit.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit>
     <testsuites>
-        <testsuite name="Injector Test Suite">
-            <directory suffix="Test.php">./test</directory>
+        <testsuite name="Unit">
+            <directory suffix="Test.php">./test/Unit</directory>
         </testsuite>
     </testsuites>
     <filter>
