+++ test/Unit/Dependency/DependencyFinderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Injector\Test\Dependency;
+namespace Injector\Test\Unit\Dependency;
 
 use Injector\Binding\BindingMap;
 use Injector\Dependency\AutoWireDependency;
